requirements.txt
azure-core
azure-identity
azure-storage-blob
microsoft-bing-websearch
playwright
python-dotenv
tldextract